- fix build_ext bug - add wheels script cases for mac wheels

+++ build_tools/python_deploy/build_macos_packages.sh
@@ -22,7 +22,7 @@ this_dir="$(cd "$(dirname "$0")" && pwd)"
 repo_root="$(cd "$this_dir"/../../ && pwd)"
 python_versions="${TORCH_MLIR_PYTHON_VERSIONS:-3.9 3.10 3.11}"
 output_dir="${output_dir:-${this_dir}/wheelhouse}"
-packages="${packages:-torch-mlir}"
+packages="${packages:-torch-mlir torch-mlir-no-jit-importer}"
 
 PKG_VER_FILE="${repo_root}"/torch_mlir_package_version ; [ -f "$PKG_VER_FILE" ] && . "$PKG_VER_FILE"
 export TORCH_MLIR_PYTHON_PACKAGE_VERSION="${TORCH_MLIR_PYTHON_PACKAGE_VERSION:-0.0.1}"
@@ -61,6 +61,11 @@ function run() {
           build_torch_mlir torch_mlir "$python_version"
           run_audit_wheel torch_mlir "$python_version"
           ;;
+        torch-mlir-no-jit-importer)
+          clean_wheels torch_mlir_no_jit "$python_version"
+          build_torch_mlir_no_jit torch_mlir_no_jit "$python_version"
+          run_audit_wheel_no_jit torch_mlir_no_jit "$python_version"
+          ;;
         *)
           echo "Unrecognized package '$package'"
           exit 1
@@ -88,6 +93,25 @@ function build_torch_mlir() {
   rm -rf "$output_dir"/build_venv
 }
 
+function build_torch_mlir_no_jit() {
+  local wheel_basename="$1"
+  local python_version="$2"
+  rm -rf "$output_dir"/build_venv
+  python"${python_version}" -m venv "$output_dir"/build_venv
+  source "$output_dir"/build_venv/bin/activate
+  python"${python_version}" -m pip install -U pip delocate
+  python"${python_version}" -m pip install -r "$repo_root"/build-requirements.txt
+  CMAKE_GENERATOR=Ninja \
+  TORCH_MLIR_PYTHON_PACKAGE_VERSION=${TORCH_MLIR_PYTHON_PACKAGE_VERSION} \
+  MACOSX_DEPLOYMENT_TARGET=$MACOSX_DEPLOYMENT_TARGET \
+  CMAKE_OSX_ARCHITECTURES=$CMAKE_OSX_ARCHITECTURES \
+  TORCH_MLIR_ENABLE_JIT_IR_IMPORTER=0 \
+  TORCH_MLIR_ENABLE_ONLY_MLIR_PYTHON_BINDINGS=1 \
+  python"${python_version}" -m pip wheel -v -w "$output_dir" "$repo_root"
+  deactivate
+  rm -rf "$output_dir"/build_venv
+}
+
 function clean_wheels() {
   local wheel_basename="$1"
   local python_version="$2"
@@ -111,6 +135,26 @@ function run_audit_wheel() {
     python"${python_version}" -m pip install -r "$repo_root"/pytorch-requirements.txt --extra-index-url https://download.pytorch.org/whl/nightly/cpu
     python"${python_version}" -m pip install -r "$repo_root"/build-requirements.txt
     python"${python_version}" -m pip install "$generic_wheel" --extra-index-url https://download.pytorch.org/whl/nightly/cpu
+    DYLD_LIBRARY_PATH="$output_dir"/test_venv/lib/python"${python_version}"/site-packages/torch/lib delocate-wheel -v "$generic_wheel"
+    deactivate
+    rm -rf "$output_dir"/test_venv
+  fi
+}
+
+function run_audit_wheel_no_jit() {
+  set +x
+  local wheel_basename="$1"
+  local python_version="$2"
+  generic_wheel=$(ls "$output_dir"/"${wheel_basename}"-* | grep "${python_version//./}")
+  echo "Looking for $generic_wheel"
+  if [ -f "$generic_wheel" ]; then
+    echo "$generic_wheel found. Delocating it.."
+    rm -rf "$output_dir"/test_venv
+    python"${python_version}" -m venv "$output_dir"/test_venv
+    source "$output_dir"/test_venv/bin/activate
+    python"${python_version}" -m pip install -U pip delocate
+    python"${python_version}" -m pip install -r "$repo_root"/build-requirements.txt
+    python"${python_version}" -m pip install "$generic_wheel"
     DYLD_LIBRARY_PATH="$output_dir"/test_venv/lib/python"${python_version}"/site-packages/torch/lib delocate-wheel -v "$generic_wheel"
     deactivate
     rm -rf "$output_dir"/test_venv
